requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.4
numpy==1.26.2
httpx==0.25.0
openai==1.51.0
python-dotenv==1.0.0
asyncio==3.4.3
gunicorn==21.2.0
requests==2.31.0
